mcp-azure-devops/requirements.txt
# MCP framework
mcp>=1.0.0

# Azure SDKs
azure-identity>=1.15.0
azure-mgmt-containerservice>=30.0.0
azure-mgmt-monitor>=6.0.0
azure-mgmt-costmanagement>=4.0.0

# Kubernetes Python client
kubernetes>=29.0.0

# HTTP client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
uvicorn>=0.29.0     # optional — for HTTP transport mode

# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.23.0
